% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_gmm.R
\name{acf_gmm_cd}
\alias{acf_gmm_cd}
\title{Second Stage GMM}
\usage{
acf_gmm_cd(param, data, instruments = c("const", "l_lag", "k", "phi_lag"))
}
\arguments{
\item{param}{numeric}

\item{data}{data frame}

\item{instruments}{character}
}
\value{
numeric
}
\description{
Evaluate the moment conditions of a Cobb-Douglas production function.
}
