% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_gmm.R
\name{acf_gmm_tl}
\alias{acf_gmm_tl}
\title{Second Stage GMM}
\usage{
acf_gmm_tl(
  param,
  data,
  instruments = c("const", "l_lag", "k", "ll_lag", "kk", "l_lag_k", "phi_lag")
)
}
\arguments{
\item{param}{numeric}

\item{data}{data frame}

\item{instruments}{character}
}
\value{
numeric
}
\description{
Evaluate the moment conditions of a Translog production function.
}
