% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_go_gmm.R
\name{acf_go_gmm_tl}
\alias{acf_go_gmm_tl}
\title{Second Stage GMM (Gross Output)}
\usage{
acf_go_gmm_tl(
  param,
  data,
  instruments = c("const", "l_lag", "k", "m_lag", "ll_lag", "kk", "mm_lag", "l_lag_k",
    "lm_lag", "k_m_lag", "l_lag_k_m_lag", "phi_lag")
)
}
\arguments{
\item{param}{numeric}

\item{data}{data frame}

\item{instruments}{character}
}
\value{
numeric
}
\description{
Evaluate the moment conditions of a gross output translog production function.
}
