\name{as.list.matrix}
\alias{as.list.matrix}
\title{Transform a matrix to a list}
\usage{
  \method{as.list}{matrix} (x, byrow = TRUE, ...)
}
\arguments{
  \item{x}{a matrix}

  \item{byrow}{each element of the list will correspond to
  each row (\code{byrow = TRUE}) or each column
  (\code{byrow = FALSE}) of the matrix}

  \item{...}{not used}
}
\value{
  a list
}
\description{
  Transform a matrix to a list
}
\author{
  David Hajage \email{dhajage@gmail.com}
}
\keyword{list}

