% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_gmm.R
\name{dlw_gmm_cd}
\alias{dlw_gmm_cd}
\title{Second Stage GMM (as in DLW)}
\usage{
dlw_gmm_cd(param, data, instruments = c("l_lag", "k"))
}
\arguments{
\item{param}{numeric}

\item{data}{data frame}

\item{instruments}{character}
}
\value{
numeric
}
\description{
Evaluate the moment conditions of a Cobb-Douglas production function.
}
