% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_gmm.R
\name{dlw_gmm_tl}
\alias{dlw_gmm_tl}
\title{Second Stage GMM (as in DLW)}
\usage{
dlw_gmm_tl(
  param,
  data,
  instruments = c("l_lag", "k", "ll_lag", "kk", "l_lag_k")
)
}
\arguments{
\item{param}{numeric}

\item{data}{data frame}

\item{instruments}{character}
}
\value{
numeric
}
\description{
Evaluate the moment conditions of a Translog production function.
}
