% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_go_gmm.R
\name{dlw_go_gmm_cd}
\alias{dlw_go_gmm_cd}
\title{Second Stage GMM (Gross Output), as in DLW/DLS}
\usage{
dlw_go_gmm_cd(
  param,
  data,
  instruments = c("l", "k", "m_lag"),
  controls = c("log_w")
)
}
\arguments{
\item{param}{numeric}

\item{data}{data frame}

\item{instruments}{character}

\item{controls}{character}
}
\value{
numeric
}
\description{
Evaluate the moment conditions of a gross output Cobb-Douglas production function.
}
