\docType{package}
\name{markup}
\alias{markup}
\alias{markup-package}
\title{markup}
\usage{
  markup()
}
\arguments{
  \item{x}{x}

  \item{include.rownames}{if TRUE, appears and strong}

  \item{include.colnames}{if TRUE, appears and strong (no
  more header argument)}

  \item{rownames}{rownames}

  \item{colnames}{colnames}

  \item{italic}{logical vector or matrix with columns/cells
  that must be italic}

  \item{strong}{logical vector or matrix with columns/cells
  that must be strong}

  \item{caption}{(and no more caption.level)}

  \item{align}{or justify (and no more valign)}

  \item{format}{format}

  \item{digits}{digits}

  \item{decimal.mark}{decimal.mark}

  \item{na.print}{na.print}

  \item{tgroup}{(appears strong and aligned, no more style,
  align... options for them)}

  \item{n.tgroup}{n.tgroup}

  \item{bgroup}{bgroup}

  \item{n.bgroup}{n.bgroup}

  \item{lgroup}{lgroup}

  \item{n.lgroup}{n.lgroup}

  \item{rgroup}{rgroup}

  \item{n.rgroup}{n.rgroup}

  \item{...}{...}
}
\value{
  a markup object
}
\description{
  markup

  Idea for the main function
}
\author{
  David Hajage \email{dhajage@gmail.com}
}

