\name{plim}
\alias{plim}
\title{format p values}
\usage{
  plim(p, digits = 2, drop0trailing = FALSE)
}
\arguments{
  \item{p}{p values}

  \item{digits}{number of digits}

  \item{drop0trailing}{see \code{?formatC}}
}
\value{
  A character vector (formated p values)
}
\description{
  format p values
}
\author{
  David Hajage
}

