% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{acf}
\alias{acf}
\title{Title}
\usage{
acf(initial_values, data, type = c("cd", "tl"), ...)
}
\arguments{
\item{initial_values}{named numeric}

\item{data}{data frame}

\item{type}{string specifying the type of production function to be estimated. Currently, only Cobb-Douglas is supported.}

\item{...}{additional arguments to be passed to \code{optim()}}
}
\value{
list
}
\description{
Title
}
