% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{acf}
\alias{acf}
\title{Title}
\usage{
acf(initial_values, data, ...)
}
\arguments{
\item{initial_values}{named numeric}

\item{data}{data frame}

\item{...}{additional arguments to be passed to \code{optim()}}
}
\value{
list
}
\description{
Title
}
